보내기

+++ server/PacketQueue.hpp
@@ -13,7 +13,7 @@ public:
 	void pushPacket( const Packet& packet );
 	void dispatch( );
 	void addObject( Object* obj );
-	void removeObject(std::uint8_t networkId);
+	void removeObject(std::uint16_t networkId);
 
 	Object* getObject( std::uint16_t id ) {
 		return networkIdToObject[ id ];
@@ -23,8 +23,8 @@ private:
 	std::queue<Packet> packetQueue_;
 	std::mutex queueMtx_;
 
-	std::unordered_map<int, Object*> networkIdToObject;
-	std::unordered_map<Object*, int> objectToNetworkId;
+	std::unordered_map<uint16_t, Object*> networkIdToObject;
+	std::unordered_map<Object*, uint16_t> objectToNetworkId;
 };
 
 #endif // PACKETQUEUE_HPP
